+++ src/fe/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "confluence-global-page-custom-ui-static",
+  "name": "doculink",
   "version": "0.1.37",
   "private": true,
   "homepage": ".",
